zookeeper distributed pool number

+++ src/main/java/pl/forcode/tinyurlservice/idGenerator/IdsPool.java
@@ -2,6 +2,6 @@ package pl.forcode.tinyurlservice.idGenerator;
 
 import java.util.Collection;
 
-public record IdsPool(int poolNumber, Collection<String> ids) {
+public record IdsPool(long poolNumber, Collection<String> ids) {
 
 }
